refactor(tests): extract shared git-fixture.js and tag real-repo leg as integration (refs #3643)
Extract the throwaway-git-repo helper from tests/baselines/git-base.test.js into
tests/fixtures/git-fixture.js — optimized with `git init -q -b main` and inline
`-c user.email / user.name / commit.gpgsign=false` flags on the commit call, replacing
three separate `git config` subprocess round-trips.

Move the single "real git" acceptance test into tests/baselines/git-base.integration.test.js
so it is picked up by the `tests/**/*.integration.test.js` glob in INTEGRATION_INCLUDE and
excluded from `test:quick`. The five mock-based unit tests remain in git-base.test.js on
the quick tier.
